Types for Checkbox component
